Restructure project/job/workflow/dataset endpoint files
